ttf_overlap_remover を置き換え
