+++ font_rasterizer/Cargo.toml
@@ -29,7 +29,7 @@ usvg = { workspace = true }
 bezier_converter = { git = "https://github.com/mitoma/sandbox", branch = "main" }
 font_collector = { path = "../font_collector" }
 text_buffer = { path = "../text_buffer" }
-ttf_overlap_remover2 = { path = "../ttf_overlap_remover2" }
+ttf_overlap_remover = { path = "../ttf_overlap_remover" }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 console_error_panic_hook = { workspace = true }
