+++ ttf_overlap_remover/src/outline_builder.rs
@@ -1,13 +1,12 @@
-use std::f32;
-
 use rustybuzz::ttf_parser::OutlineBuilder;
-use tiny_skia_path::{Path, PathBuilder, Point};
+use tiny_skia_path::{Path, PathBuilder};
 
-use crate::{
-    path_segment::{Cubic, Line, PathSegment, Quadratic},
-    remove_overlap, remove_path_overlap,
-};
+use crate::remove_path_overlap;
 
+/// オーバーラップ除去機能付き OutlineBuilder。
+///
+/// `rustybuzz::Face::outline_glyph()` に渡して使用する。
+/// グリフのアウトラインを内部に蓄積し、`removed_paths()` で重複除去後のパスを取得できる。
 #[derive(Debug)]
 pub struct OverlapRemoveOutlineBuilder {
     builder: Option<PathBuilder>,
@@ -28,51 +27,49 @@ impl OverlapRemoveOutlineBuilder {
         }
     }
 
+    /// 元のパス（重複除去前）を取得する
     pub fn paths(&self) -> Vec<Path> {
         self.paths.clone()
     }
 
+    /// 重複除去後のパスを取得する
     pub fn removed_paths(&self) -> Vec<Path> {
         remove_path_overlap(self.paths.clone())
     }
 
+    /// 重複除去後のパスを OutlineBuilder にフィードバックする
     pub fn outline<T>(&self, builder: &mut T)
     where
         T: OutlineBuilder,
     {
-        let removed_paths = remove_overlap(self.paths.clone());
-
-        removed_paths.iter().for_each(|path| {
-            let segments_len = path.segments.len();
-            path.segments.iter().enumerate().for_each(|(i, segment)| {
-                if i == 0 {
-                    let Point { x, y } = segment.endpoints().0;
-                    builder.move_to(x, y);
-                }
-
+        let removed = self.removed_paths();
+        for path in &removed {
+            let mut first = true;
+            for segment in path.segments() {
                 match segment {
-                    PathSegment::Line(Line { to, .. }) => {
-                        builder.line_to(to.x, to.y);
+                    tiny_skia_path::PathSegment::MoveTo(p) => {
+                        builder.move_to(p.x, p.y);
+                        first = false;
                     }
-                    PathSegment::Quadratic(Quadratic { control, to, .. }) => {
-                        builder.quad_to(control.x, control.y, to.x, to.y);
+                    tiny_skia_path::PathSegment::LineTo(p) => {
+                        if first {
+                            builder.move_to(p.x, p.y);
+                            first = false;
+                        }
+                        builder.line_to(p.x, p.y);
                     }
-                    PathSegment::Cubic(Cubic {
-                        control1,
-                        control2,
-                        to,
-                        ..
-                    }) => {
-                        builder
-                            .curve_to(control1.x, control1.y, control2.x, control2.y, to.x, to.y);
+                    tiny_skia_path::PathSegment::QuadTo(c, p) => {
+                        builder.quad_to(c.x, c.y, p.x, p.y);
+                    }
+                    tiny_skia_path::PathSegment::CubicTo(c1, c2, p) => {
+                        builder.curve_to(c1.x, c1.y, c2.x, c2.y, p.x, p.y);
+                    }
+                    tiny_skia_path::PathSegment::Close => {
+                        builder.close();
                     }
                 }
-
-                if i == segments_len - 1 {
-                    builder.close();
-                }
-            });
-        });
+            }
+        }
     }
 }
 
@@ -99,6 +96,8 @@ impl OutlineBuilder for OverlapRemoveOutlineBuilder {
     fn close(&mut self) {
         let mut builder = self.builder.replace(PathBuilder::new()).unwrap();
         builder.close();
-        self.paths.push(builder.finish().unwrap());
+        if let Some(path) = builder.finish() {
+            self.paths.push(path);
+        }
     }
 }
