+++ ttf_overlap_remover/src/winding.rs
